# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewPeace BookStore API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/tag/search": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "搜尋標籤",
                "operationId": "3367bd73de6668003ad4d2168b9d9856",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "description": "關鍵字",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "data": [
                                        {
                                            "id": 1,
                                            "name": "tag-1",
                                            "created_at": "2021-11-12T15:15:10.000000Z",
                                            "updated_at": "2021-11-12T15:15:10.000000Z"
                                        },
                                        {
                                            "id": 2,
                                            "name": "tag-2",
                                            "created_at": "2021-11-12T15:15:10.000000Z",
                                            "updated_at": "2021-11-12T15:15:10.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tag/{id}/items": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "取得標籤相關商品",
                "operationId": "74971200e1745d7947da42dfbbdf67c5",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "data": [
                                        {
                                            "id": 1,
                                            "owner": 1,
                                            "category": 1,
                                            "name": "Book Name",
                                            "ISBN": "3333-1111-2222-1234",
                                            "price": 100,
                                            "quantity": 1,
                                            "created_at": "2021-11-12T15:15:10.000000Z",
                                            "updated_at": "2021-11-12T15:15:10.000000Z"
                                        },
                                        {
                                            "id": 2,
                                            "owner": 2,
                                            "category": 2,
                                            "name": "Book Name2",
                                            "ISBN": "7777-1666-5552-4321",
                                            "price": 800,
                                            "quantity": 2,
                                            "created_at": "2021-11-12T15:15:10.000000Z",
                                            "updated_at": "2021-11-12T15:15:10.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tag/create": {
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "新增標籤",
                "operationId": "205a0609f82ae7f3ffb78f4f4e1cb95e",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "標籤名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "data": {
                                        "id": 1,
                                        "name": "tag-1",
                                        "created_at": "2021-11-12T15:15:10.000000Z",
                                        "updated_at": "2021-11-12T15:15:10.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "失敗(名稱重複)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "duplicate tag name"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "失敗(請求格式錯誤)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "error Input"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tag/{id}/delete": {
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "刪除標籤",
                "operationId": "e4fa78aa1c39508b0c48643aa3ea0253",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "失敗(標籤不存在)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "tag not found"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}